% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_col.R
\name{add_col_in_nest_byGRP}
\alias{add_col_in_nest_byGRP}
\title{add variable}
\usage{
add_col_in_nest_byGRP(df, group_vars, mutate_vars)
}
\arguments{
\item{df}{A dataframe}

\item{mutate_vars}{The variables you want to add to the dataframe}

\item{group_var}{The variable to be grouped}
}
\description{
Add variables as you want/needed
}
