% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_col.R
\name{add_var_pct}
\alias{add_var_pct}
\title{add variable}
\usage{
add_var_pct(df, var)
}
\arguments{
\item{df}{A dataframe}

\item{var}{Variable to be calculated}
}
\description{
Add variables as you want/needed
}
