% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{arrange_double_nested}
\alias{arrange_double_nested}
\title{Arrange rows by variables}
\usage{
arrange_double_nested(df, nest_cols, ...)
}
\arguments{
\item{df}{A nested dataframe}

\item{nest_cols}{Name of columns to nest in existing list-column}

\item{...}{Comma separated list of unquoted variable names}
}
\description{
Arrange rows by variables in double nested dataframe
}
