% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_var.R
\name{arrange_groupVar}
\alias{arrange_groupVar}
\title{arrange 
arrange order by certain variable}
\usage{
arrange_groupVar(df, group_var, ...)
}
\arguments{
\item{df}{A nested dataframe}

\item{group_var}{The varaible to be grouped}
}
\description{
arrange 
arrange order by certain variable
}
