% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_var.R
\name{arrange_in_nest}
\alias{arrange_in_nest}
\title{arrange in nested dataframe}
\usage{
arrange_in_nest(df, group_var, ...)
}
\arguments{
\item{df}{A nested dataframe}

\item{group_var}{The variable to be grouped}

\item{...}{Variables or functions}
}
\description{
arrange in nested dataframe
}
