% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{arrange_nested}
\alias{arrange_nested}
\title{Arrange rows by variables}
\usage{
arrange_nested(df, ...)
}
\arguments{
\item{df}{A nested dataframe}

\item{...}{Comma separated list of unquoted variable names}
}
\description{
Arrange rows by variables in nested dataframe
}
