% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_var.R
\name{arrange_var}
\alias{arrange_var}
\title{arrange
arrange order by certain variable}
\usage{
arrange_var(df, ...)
}
\arguments{
\item{df}{A nested dataframe}
}
\description{
arrange
arrange order by certain variable
}
