% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal-variables.R
\name{calcu_day}
\alias{calcu_day}
\title{Calculate day variable.}
\usage{
calcu_day(df)
}
\arguments{
\item{df}{A dataframe}
}
\description{
Calculate unique day of a week of each tract, and store the result in a list.
}
