% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-scoring.R
\name{calcu_daytimes}
\alias{calcu_daytimes}
\title{Calculate daytime variable.}
\usage{
calcu_daytimes(data)
}
\arguments{
\item{data}{A dataframe from a nested filtered dataframe}
}
\description{
Calculate the percentage of tweets sent on work time and night time of each tract, throw away tracts that percentage 
of work time is larger than that of night time, only keep percentage of night time and store the result in a list.
}
