% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal-variables.R
\name{calcu_month}
\alias{calcu_month}
\title{Calculate month variable.}
\usage{
calcu_month(df)
}
\arguments{
\item{df}{A dataframe}
}
\description{
Calculate unique month of a year of each tract, and store the result in a list.
}
