% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-scoring.R
\name{calcu_satmor}
\alias{calcu_satmor}
\title{Calculate Sat morning variable.}
\usage{
calcu_satmor(data)
}
\arguments{
\item{data}{A dataframe from a nested filtered dataframe}
}
\description{
Calculate the percentage of tweets sent on Sat morning of each tract, filter tweets all sent in the morning first, 
calculate the percentage of tweets sent all days of a week and only keep the percentage that sent on Sat and store 
the result in a list.
}
