% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-scoring.R
\name{calcu_week}
\alias{calcu_week}
\title{Calculate week variable.}
\usage{
calcu_week(data)
}
\arguments{
\item{data}{A dataframe from a nested filtered dataframe}
}
\description{
Calculate the percentage of weekday and weekend of each tract, throw away tract that only has tweets sent on 
weekday and keep weekend percent value and store the result in a list.
}
