% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_timestamp.R
\name{derive_timestamp}
\alias{derive_timestamp}
\title{Basic variables}
\usage{
derive_timestamp(df, timestamp)
}
\arguments{
\item{df}{A dataframe with columns for the user id, location, timestamp}

\item{timestamp}{Name of timestamp column. Should be POSIXct}
}
\description{
Add basic variables derive from timestamp
}
