% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_neighbors}
\alias{df_neighbors}
\title{Neighbours for locations}
\format{
A nested dataframe of 1,942 rows and 2 variables:
\describe{
\item{grid_id}{unique identifier for each location}
\item{neighbor}{computed neighbours of the location with TOUCH relation}
}
}
\usage{
df_neighbors
}
\description{
Spatial neighbours for locations, where a neighbour has
at least one line in common, but its interior does not intersect with the location
}
\keyword{datasets}
