% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_timestamp.R
\name{enrich_timestamp}
\alias{enrich_timestamp}
\title{Derive new variables}
\usage{
enrich_timestamp(df, timestamp, tz = "Asia/Singapore")
}
\arguments{
\item{df}{A nested dataframe}

\item{timestamp}{Name of timestamp column. Should be POSIXct}

\item{tz}{Time zone of timestamp}
}
\description{
Add basic needed variables derived from timestamp
}
