% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_home.R
\name{extract_home}
\alias{extract_home}
\title{extract location}
\usage{
extract_home(df, ..., show_n_home = 1, keep_score = F)
}
\arguments{
\item{df}{A nested dataframe by user}

\item{show_n_home}{Number of homes to be shown}

\item{keep_score}{Choice to keep score or not}
}
\description{
Extract most likely home location of each user
}
