% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_loc.R
\name{extract_loc}
\alias{extract_loc}
\title{extract location}
\usage{
extract_loc(
  df,
  score_var,
  user = "u_id",
  location = "grid_id",
  show_n_loc = 1,
  keep_score = F
)
}
\arguments{
\item{df}{A nested dataframe by user}

\item{score_var}{Name of column that holds weighted score for each user}

\item{user}{Name of column that holds unique identifier for each user}

\item{location}{Name of column that holds unique identifier for each location}

\item{show_n_loc}{Number of homes to be shown}

\item{keep_score}{Choice to keep score or not}
}
\description{
Extract most likely home location of each user
}
