% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_location.R
\name{extract_location}
\alias{extract_location}
\title{Extract identified home locations for users}
\usage{
extract_location(
  df,
  user = "u_id",
  location = "loc_id",
  show_n_loc = 1,
  keep_score = F,
  ...
)
}
\arguments{
\item{df}{A nested dataframe by user}

\item{user}{Name of column that holds unique identifier for each user}

\item{location}{Name of column that holds unique identifier for each location}

\item{show_n_loc}{Number of homes to be shown}

\item{keep_score}{Choice to keep score or not}

\item{score_var}{Name of column that holds weighted score for each user}
}
\description{
Extract the most likely home location for each user based on choosed column value
}
