% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_var.R
\name{filter_in_nest}
\alias{filter_in_nest}
\title{Filter
Keep only users that meet certain preconditions}
\usage{
filter_in_nest(df, ...)
}
\arguments{
\item{df}{A nested dataframe grouped by user}

\item{...}{Variables or functions}
}
\description{
Filter
Keep only users that meet certain preconditions
}
