% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_verbose}
\alias{filter_verbose}
\title{Return rows with matching conditions}
\usage{
filter_verbose(df, user = "u_id", ...)
}
\arguments{
\item{df}{A dataframe}

\item{user}{Name of column that holds unique identifier for each user}

\item{...}{Logical predicates defined in terms of the variables in df. Only rows match conditions are kept.}
}
\description{
Filter finds rows where conditions are true.
}
