% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_var.R
\name{grpSumm_in_nest}
\alias{grpSumm_in_nest}
\title{Computed variables by group}
\usage{
grpSumm_in_nest(df, nest_cols, summary_vars)
}
\arguments{
\item{df}{A nested dataframe}

\item{nest_cols}{Variables to be nested}

\item{summary_vars}{Summarise expression}
}
\description{
Add needed variables by group
}
