% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home_extraction.R
\name{home_extract}
\alias{home_extract}
\title{Extract home by temporal variable score.}
\usage{
home_extract(df)
}
\arguments{
\item{df}{A dataframe}
}
\description{
Estimate a user's home location according to the socre calculated by temporal varialbes.
}
