% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home-filtering.R
\name{home_filtering}
\alias{home_filtering}
\title{Filter scared data.}
\usage{
home_filtering(df_data)
}
\arguments{
\item{df_data}{A dataframe with census tract}
}
\description{
Kepp users have more than 20 tweets totally, and have more than 10 tweets, more than 10 unique days and more than 8 hours tweets in a tract record.
}
