% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{identify_home}
\alias{identify_home}
\title{one-button identify home location}
\usage{
identify_home(df, user = "u_id", timestamp = "created_at",
  location = "grid_id", recipe = "homelocator")
}
\arguments{
\item{df}{A dataframe with columns for the user id, location, timestamp}

\item{user}{Name of column that holds unique identifier for each user}

\item{timestamp}{Name of timestamp column. Should be POSIXct}

\item{location}{Name of column that holds unique identifier for each location}

\item{recipe}{Different methods to identify home locations}
}
\description{
TIdentify home location with chose recipe
}
