% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{identify_loc}
\alias{identify_loc}
\title{one-button identify home location}
\usage{
identify_loc(
  df,
  user = "u_id",
  timestamp = "created_at",
  location = "grid_id",
  recipe = "HLC",
  show_n_loc = 1,
  rm_pct = 0.01
)
}
\arguments{
\item{df}{A dataframe with columns for the user id, location, timestamp}

\item{user}{Name of column that holds unique identifier for each user}

\item{timestamp}{Name of timestamp column. Should be POSIXct}

\item{location}{Name of column that holds unique identifier for each location}

\item{recipe}{Different methods to identify home locations}

\item{show_n_home}{Number of potential homes to be shown}
}
\description{
Identify home location with chose recipe
}
