% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_location.R
\name{identify_location}
\alias{identify_location}
\title{Identify home locations for users}
\usage{
identify_location(
  df,
  user = "u_id",
  timestamp = "created_at",
  location = "loc_id",
  tz = "Asia/Singapore",
  recipe,
  show_n_loc = 1,
  keep_score = F
)
}
\arguments{
\item{df}{A dataframe with columns for the user id, location, timestamp}

\item{user}{Name of column that holds unique identifier for each user}

\item{timestamp}{Name of timestamp column. Should be POSIXct}

\item{location}{Name of column that holds unique identifier for each location}

\item{tz}{A character string containing the time zone to convert to and it should be recognized in R}

\item{recipe}{Embeded algorithms to identify the most possible home locations for users}

\item{show_n_loc}{Number of potential homes to extract}

\item{keep_score}{Option to keep or remove calculated result/score per user per location}
}
\description{
This function infers the most possible home locations for users with "recipe".
The "recipe" can be either embeded "recipe" - HMLC, OSNA, APDM, FREQ, or self created "recipe".
}
