% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate_double_nested}
\alias{mutate_double_nested}
\title{Add new variable}
\usage{
mutate_double_nested(df, nest_cols, ...)
}
\arguments{
\item{df}{A dataframe}

\item{nest_cols}{A selection of columns to nest in existing list-column}

\item{...}{Name-value pairs of functions}
}
\description{
Add new variable in double nested dataset
}
