% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate_nested}
\alias{mutate_nested}
\title{Add new variables in list-column}
\usage{
mutate_nested(df, ...)
}
\arguments{
\item{df}{A nested dataframe}

\item{...}{Name-value pairs of expressions}
}
\description{
Add new variables and preserves existing from list-column
}
