% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate_verbose}
\alias{mutate_verbose}
\title{Add new variable}
\usage{
mutate_verbose(df, ...)
}
\arguments{
\item{df}{A dataframe}

\item{...}{Name-value pairs of expressions}
}
\description{
Add new variable and perserves existing to dataframe
}
