% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_dataframe.R
\name{nest_by_mulGps}
\alias{nest_by_mulGps}
\title{Nest dataframe}
\usage{
nest_by_mulGps(df, group_vars)
}
\arguments{
\item{df}{A dataframe}

\item{group_vars}{A list contains variables to be grouped}
}
\description{
Nest dataframe by multiple varibles
}
