% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_dataframe.R
\name{nest_by_sglGp}
\alias{nest_by_sglGp}
\title{Nest dataframe}
\usage{
nest_by_sglGp(df, group_var)
}
\arguments{
\item{df}{A dataframe}

\item{group_var}{variable to be grouped}
}
\description{
Nest dataframe by variable
}
