% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_dataframe.R
\name{nest_dataframe}
\alias{nest_dataframe}
\title{Nest dataframe}
\usage{
nest_dataframe(df, group_var)
}
\arguments{
\item{df}{A dataframe}

\item{group_var}{variable to be grouped by}
}
\description{
Nest dataframe by variable
}
