% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{nest_nested}
\alias{nest_nested}
\title{Nest nested dataframe}
\usage{
nest_nested(df, ...)
}
\arguments{
\item{df}{A nested dataframe}

\item{...}{A selection of columns.}
}
\description{
Double nesting creates a list-column of nested dataframe
}
