% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{nest_verbose}
\alias{nest_verbose}
\title{Nest dataframe}
\usage{
nest_verbose(df, ...)
}
\arguments{
\item{df}{A dataframe}

\item{...}{A selection of columns.}
}
\description{
Nesting creates a list-column of dataframe
}
