% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal-variables.R
\name{para_value}
\alias{para_value}
\title{Gather temporal variables' results.}
\usage{
para_value(home_filter)
}
\arguments{
\item{df}{A nested dataframe}
}
\description{
Gather resutls get from funciton: calcu_week, calcu_satmor, calcu_daytimes, calcu_day, calcu_day, calcu_month 
and delete data dataframe
}
