% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{prop_factor_nested}
\alias{prop_factor_nested}
\title{Calculate the proportion of categories for a variable within a nested dataframe}
\usage{
prop_factor_nested(df, ...)
}
\arguments{
\item{df}{A nested dataframe}

\item{var}{Name of column to calculate}
}
\description{
Calculate the proportion of categories for a variable within a nested dataframe and convert each categories to a new variable adding to the dataframe
}
