% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_location.R
\name{recipe_FREQ}
\alias{recipe_FREQ}
\title{recipe: frequency - FREQ}
\usage{
recipe_FREQ(
  df,
  user = "u_id",
  timestamp = "created_at",
  location = "loc_id",
  show_n_loc,
  keep_score = F
)
}
\description{
recipe: frequency - FREQ
}
