% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_location.R
\name{recipe_HMLC}
\alias{recipe_HMLC}
\title{recipe: homelocator - HMLC}
\usage{
recipe_HMLC(
  df,
  user = "u_id",
  timestamp = "created_at",
  location = "loc_id",
  show_n_loc,
  keep_original_vars = F,
  keep_score = F
)
}
\description{
recipe: homelocator - HMLC
}
