% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_location.R
\name{recipe_OSNA}
\alias{recipe_OSNA}
\title{recipe: Online Social Networks Activity - OSNA}
\usage{
recipe_OSNA(
  df,
  user = "u_id",
  timestamp = "created_at",
  location = "loc_id",
  show_n_loc,
  keep_score = F
)
}
\description{
recipe: Online Social Networks Activity - OSNA
}
