% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_top_users.R
\name{remove_top_users}
\alias{remove_top_users}
\title{Remove users}
\usage{
remove_top_users(df, user = "u_id", counts = "n_points", topNpct_user = 1)
}
\arguments{
\item{df}{A dataframe with columns for the user id, counts point per user}

\item{user}{Name of column that holds unique identifier for each user}

\item{counts}{Name of column that holds the data points frequency for each user}

\item{topNpct_user}{A decimal number that represent the certain percentage of users to remove}
}
\description{
Remove certain percent of top active users based on the total number of data points
}
