% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_nested}
\alias{score_nested}
\title{score of variables}
\usage{
score_nested(
  df,
  user = "u_id",
  location = "loc_id",
  keep_original_vars = F,
  ...
)
}
\arguments{
\item{df}{A nested dataframe by user}

\item{user}{Name of column that holds unique identifier for each user}

\item{location}{Name of column that holds unique identifier for each location}

\item{...}{Name-value pairs of expression

Can we replace this with just mutate_nested? And potentially add transmute_nested counterpart?}

\item{keep_ori_vars}{Option to keep or drop original varialbes}
}
\description{
Add score of each variables
}
