% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal-variables.R
\name{scoring}
\alias{scoring}
\title{Give a score to each tract for each user.}
\usage{
scoring(df)
}
\arguments{
\item{df}{A list}
}
\description{
Give a 0-1 range score to each variable and add up them to get a final score for each tract, order the tracts 
according to the score.
}
