% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_var.R
\name{spread2_add_missing}
\alias{spread2_add_missing}
\title{spread variable}
\usage{
spread2_add_missing(df, full_col)
}
\arguments{
\item{df}{A nested dataframe}

\item{full_col}{Full columns in nested data}
}
\description{
Spread variables to tidy dataframe
}
