% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_var.R
\name{spread2_in_nest}
\alias{spread2_in_nest}
\title{spread variable}
\usage{
spread2_in_nest(df, key_var, value_var)
}
\arguments{
\item{df}{A dataframe}

\item{key_var}{Key variable to be spreaded}

\item{value_var}{Value variable to be spreaded}
}
\description{
Spread variables to tidy dataframe
}
