% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread.R
\name{spread_nested}
\alias{spread_nested}
\title{Spread a key-value pair across multiple columns}
\usage{
spread_nested(df, key_var, value_var)
}
\arguments{
\item{df}{A nested dataframe}

\item{var_key}{Column name or position}

\item{var_value}{Column name or position}
}
\description{
Spread a key-value pair across multiple columns in nested dataframe
}
