% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_user.R
\name{sum_score}
\alias{sum_score}
\title{sum scored variables}
\usage{
sum_score(df, user = "u_id", location = "grid_id", ...)
}
\arguments{
\item{df}{A  dataframe}

\item{user}{Name of column that holds unique identifier for each user}

\item{location}{Name of column that holds unique identifier for each location}

\item{...}{Variables or functions}
}
\description{
Sum up scored variables
}
