% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_double_nested}
\alias{summarise_double_nested}
\title{Aggregate multiple values to a single value}
\usage{
summarise_double_nested(df, nest_cols, ...)
}
\arguments{
\item{df}{A nested dataframe}

\item{nest_cols}{A selection of columns to nest in existing list-column}

\item{...}{Name-value pairs of summary functions}
}
\description{
Create a list-column in existing list-column and aggregate multiple values in created list-column to a single value by name-value paired summary function
}
