% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_nested}
\alias{summarise_nested}
\title{Aggregate multiple values to a single value within a nested tibble}
\usage{
summarise_nested(df, ...)
}
\arguments{
\item{df}{A nested dataframe}

\item{...}{Name-value pairs of summary functions.}
}
\description{
Aggregate multiple values to a single value of existing table by name-value paired summary function within a nested tibble
}
