% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_sample}
\alias{test_sample}
\title{Tweets sent by 100 random users}
\format{
A dataframe of 16,300 rows and 3 variables:
\describe{
\item{u_id}{unique identifier for each user}
\item{created_at}{specific timestamp for each data point in POSIXctformat}
\item{grid_id}{unique identifier for each location}
}
}
\usage{
test_sample
}
\description{
De-identified test data includes 100 random users,
and it can be used as an example to get started with homelocator package
}
\keyword{datasets}
