% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-scoring.R
\name{to_dataframe}
\alias{to_dataframe}
\title{Convert list to dataframe.}
\usage{
to_dataframe(list_df)
}
\arguments{
\item{list_df}{A list results}
}
\description{
Convert list resutls to a dataframe and nest it by user id.
}
