% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_var.R
\name{top_n_in_nest}
\alias{top_n_in_nest}
\title{select the top entries in each group}
\usage{
top_n_in_nest(df, n = 2, wt)
}
\arguments{
\item{df}{A dataframe}

\item{n}{Number of entries to be selected}

\item{wt}{The variable to be used for ordering}
}
\description{
Select the top entries in each group, ordered by wt
}
