% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{unnest_double_nested}
\alias{unnest_double_nested}
\title{Nunest dataframe}
\usage{
unnest_double_nested(df, ...)
}
\arguments{
\item{df}{A nested dataframe}

\item{...}{Specification of columns to unnest.}
}
\description{
Unnesting makes each element of the list its own row.
}
