% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{unnest_verbose}
\alias{unnest_verbose}
\title{Unnest a list column}
\usage{
unnest_verbose(df, ...)
}
\arguments{
\item{df}{A dataframe}

\item{...}{Specification of columns to unnest.}
}
\description{
Unnesting makes each element of the list its own row.
}
