% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_dataset.R
\name{validate_dataset}
\alias{validate_dataset}
\title{Validate input dataset}
\usage{
validate_dataset(
  df,
  user = "u_id",
  timestamp = "created_at",
  location = "loc_id",
  keep_other_vars = F
)
}
\arguments{
\item{df}{A dataframe with columns for the user id, location, timestamp}

\item{user}{Name of column that holds unique identifier for each user}

\item{timestamp}{Name of column that holds specific timestamp for each data point and it should be POSIXct}

\item{location}{Name of column that holds unique identifier for each location}

\item{keep_other_vars}{Option to keep or remove other variables of the input dataset}
}
\description{
To make sure the input dataset contains all three necessary variables:
a unique identifier for the person or user
a unique identifier for the spatial locaiton for the data point
and a timestamp that relects the time the data point was created
}
