% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homelocation-filtering.R
\name{var_expand}
\alias{var_expand}
\title{Expand variables to dataset.}
\usage{
var_expand(sf_df)
}
\arguments{
\item{sf_df}{A nested dataframe}
}
\value{
A nested expanded dataframe
}
\description{
Add "year, month, day, day of week, hour of day, week, daytimes, times and group columnes to dataset. 
"year, month, day, day of week, hour of day, week, daytimes, times" are produced from timestamped variable,
and group is produced from total tweets counts according to diffrent range.
}
