% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-scoring.R
\name{calcu_month}
\alias{calcu_month}
\title{Calculate month variable.}
\usage{
calcu_month(data)
}
\arguments{
\item{data}{A dataframe from a nested filtered dataframe}
}
\description{
Calculate unique month of a year of each tract, and store the result in a list.
}
