% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-scoring.R
\name{combine_values}
\alias{combine_values}
\title{Combine all info for each user.}
\usage{
combine_values(df_filter, num)
}
\arguments{
\item{df_filter}{A nested dataframe from filtering step}

\item{num}{A number of left users}
}
\description{
Combine all info for each user including: u_id, GEOID, total_counts, counts, study_period, unique_days, unique_months, 
unique_dayofweek, unique_hours,percent_weekend, percent_satMor, percent_night, group and replace na with 0
}
