% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_timestamp.R
\name{derive_timestamp}
\alias{derive_timestamp}
\title{Derive new variables}
\usage{
derive_timestamp(df, timestamp)
}
\arguments{
\item{df}{A nested dataframe}

\item{timestamp}{Name of timestamp column. Should be POSIXct}
}
\description{
Add basic needed variables derived from timestamp
}
