% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_timestamp.R
\name{enrich_timestamp}
\alias{enrich_timestamp}
\title{Create variables}
\usage{
enrich_timestamp(df, timestamp = "created_at", tz = "Asia/Singapore")
}
\arguments{
\item{df}{A nested dataframe}

\item{timestamp}{Name of column that holds specific timestamp for each data point and it should be POSIXct}

\item{tz}{A character string containing the time zone to convert to and it should be recognized in R.}
}
\description{
Create new variables from existing timestamp
}
