% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_home.R
\name{extract_home}
\alias{extract_home}
\title{extract location}
\usage{
extract_home(df, keep_score = F, ...)
}
\arguments{
\item{df}{A nested dataframe by user}

\item{score_var}{A scored variable}
}
\description{
Extract most likely home location of each user
}
