% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_var.R
\name{filter_var}
\alias{filter_var}
\title{Filter 
Keep only users that meet certain preconditions}
\usage{
filter_var(df, filter_exp)
}
\arguments{
\item{df}{A nested dataframe grouped by user}

\item{filter_exp}{A certain condition to meet}
}
\description{
Filter 
Keep only users that meet certain preconditions
}
