% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home_extraction.R
\name{home_extract_bycounts}
\alias{home_extract_bycounts}
\title{Extract home by total tweets counts.}
\usage{
home_extract_bycounts(data)
}
\arguments{
\item{data}{A dataframe from a nested filtered dataframe}
}
\description{
Estimate a user's home location according to the total number of tweets.
}
