% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homelocation-filtering.R
\name{home_filtering}
\alias{home_filtering}
\title{Filter scared data.}
\usage{
home_filtering(df)
}
\arguments{
\item{df}{A dataframe with timestamped variable}
}
\description{
Kepp users have more than 20 tweets totally, and have more than 10 tweets, more than 10 unique days and more than 8 hours tweets in a tract record.
}
