% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-scoring.R
\name{para_value}
\alias{para_value}
\title{Gather temporal variables' results.}
\usage{
para_value(df_filter)
}
\arguments{
\item{df_filter}{A nested dataframe from filtering step}
}
\description{
Gather resutls get from funciton: calcu_week, calcu_satmor, calcu_daytimes, calcu_day, calcu_day, calcu_month 
and delete data dataframe
}
