% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{prop_factor_nested}
\alias{prop_factor_nested}
\title{Calculate relative frequency}
\usage{
prop_factor_nested(df, ...)
}
\arguments{
\item{df}{A nested dataframe}

\item{var}{Name of column to calculate}
}
\description{
Calculate relative frequency per category of the variable and spread categories to columns
}
