% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_location.R
\name{recipe_APDM}
\alias{recipe_APDM}
\title{recipe: Anchor Point Determining Model - APDM}
\usage{
recipe_APDM(
  df,
  df_neighbors,
  user = "u_id",
  timestamp = "created_at",
  location = "loc_id",
  keep_score = F
)
}
\description{
recipe: Anchor Point Determining Model - APDM
}
