% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_location.R
\name{recipe_HMLC}
\alias{recipe_HMLC}
\title{recipe: homelocator - HMLC}
\usage{
recipe_HMLC(
  df,
  user = "u_id",
  timestamp = "created_at",
  location = "loc_id",
  show_n_loc,
  keep_original_vars = F,
  keep_score = F
)
}
\arguments{
\item{df}{An enriched dataframe}

\item{user}{Name of column that holds unique identifier for each user}

\item{timestamp}{Name of timestamp column. Should be POSIXct}

\item{location}{Name of column that holds unique identifier for each location}

\item{show_n_loc}{Number of potential homes to extract}

\item{keep_original_vars}{Option to keep or remove columns other than 'user, timestamp, and location'}

\item{keep_score}{Option to keep or remove calculated result/score per user per location}
}
\description{
recipe: homelocator - HMLC
}
