% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_bots.R
\name{remove_bots}
\alias{remove_bots}
\title{remove bots}
\usage{
remove_bots(df, user = "u_id", counts = "counts_per_user",
  top_user_percent)
}
\arguments{
\item{df}{A dataframe with columns for the user id, counts point per user}

\item{user}{Name of column that holds unique identifier for each user}

\item{counts}{Number of data points per user}

\item{top_user_percent}{The percentage of top user you want to remove}
}
\description{
remove certain percentage of top users to avoid bots
}
