% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_summary}
\alias{score_summary}
\title{Sum up scores}
\usage{
score_summary(df, user = "u_id", location = "loc_id", ...)
}
\arguments{
\item{df}{A  dataframe}

\item{user}{Name of column that holds unique identifier for each user}

\item{location}{Name of column that holds unique identifier for each location}

\item{...}{A selection of columns to sum}
}
\description{
Sum up scores of variables for each location
}
