% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_user.R
\name{score_var}
\alias{score_var}
\title{score of variables}
\usage{
score_var(df, group_var, user = "u_id", location = "grid_id",
  keep_original_vars = F, ...)
}
\arguments{
\item{df}{A nested dataframe by user}

\item{group_var}{Variable to be grouped}

\item{user}{Name of column that holds unique identifier for each user}

\item{location}{Name of column that holds unique identifier for each location}
}
\description{
Add score of each variables
}
