% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-scoring.R
\name{scoring}
\alias{scoring}
\title{Give a score to each tract for each user.}
\usage{
scoring(df_result)
}
\arguments{
\item{df_result}{A dataframe combined all user info}
}
\description{
Give a 0-1 range score to each variable and add up them to get a final score for each tract, order the tracts 
according to the score.
}
