% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_var.R
\name{summarise_groupVar}
\alias{summarise_groupVar}
\title{Computed variables by group}
\usage{
summarise_groupVar(df, group_vars, summary_vars)
}
\arguments{
\item{df}{A nested dataframe}
}
\description{
Add needed variables by group
}
