% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal-variables.R
\name{to_dataframe}
\alias{to_dataframe}
\title{Convert list to dataframe.}
\usage{
to_dataframe(df)
}
\arguments{
\item{df}{A list}
}
\description{
Convert list resutls to a dataframe and nest it by user id.
}
