% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_dataset.R
\name{validate_dataset}
\alias{validate_dataset}
\title{Validate used dataset}
\usage{
validate_dataset(
  df,
  user = "u_id",
  timestamp = "created_at",
  location = "grid_id",
  keep_other_vars = F
)
}
\arguments{
\item{df}{A dataframe with columns for the user id, location, timestamp}

\item{user}{Name of column that holds unique identifier for each user}

\item{timestamp}{Name of timestamp column. Should be POSIXct}

\item{location}{Name of column that holds unique identifier for each location}

\item{keep_other_vars}{Choice to keep variables or not}
}
\description{
To make sure the used dataset contains all three necessary variables: user, location and timestamp
}
